requirements.txt
boto3==1.26.61
credit==0.0.1
pytest==7.2.1
rstr==3.2.0
selenium==4.8.0
webdriver_manager==3.8.5
Appium-Python-Client==2.8.1
pytest-bdd==6.1.1